feat(agent): add schema versioning & refactor migration to plugin pipeline
- Add SessionHeader with schema_version field for forward-compatible JSONL format

- Refactor 928-line migrate.go into extensible MigratePass pipeline (4 passes)

- Extract Sanitize pipeline in provider package for unified backfill/repair ops

- Maintain full backward compatibility with legacy session files
